Adds delete team
I also ade it where if  you type it will not auto correct team names just in case they want it spelled wrong. Also I made it where it will show the stats of the team one right when you go to the stats page. The last thing I did is when you delete a team the stats will go to zero for that team on the display. Also I fixed a few specific small bugs.